feat: Implement GUI components for presets, recording, settings, and tuner
- Added GuiPresets class for managing preset save/load/delete UI and logic.
- Introduced GuiRecording class for recording controls and waveform display.
- Created GuiSettings class for audio settings management.
- Developed GuiTuner class for chromatic tuner UI.
- Refactored PresetManager to utilize EffectFactory for effect creation.
- Removed direct effect creation methods from PresetManager.
- Enhanced recording functionality with save dialog and status messages.
- Improved UI responsiveness and aesthetics in various components.
