+++ src/audio/effect_factory.h
@@ -1,0 +1,64 @@
+#pragma once
+
+#include "audio/effect.h"
+#include <functional>
+#include <unordered_map>
+#include <string>
+
+namespace GuitarAmp {
+
+/**
+ * Registry-based effect factory.
+ * Replaces the if-chain in PresetManager::create_effect().
+ * All effect types self-register via EffectRegistrar.
+ */
+class EffectFactory {
+public:
+    using Creator = std::function<std::shared_ptr<Effect>()>;
+
+    static EffectFactory& instance() {
+        static EffectFactory factory;
+        return factory;
+    }
+
+    void register_effect(const std::string& type_name, Creator creator) {
+        creators_[type_name] = std::move(creator);
+    }
+
+    std::shared_ptr<Effect> create(const std::string& type_name) const {
+        auto it = creators_.find(type_name);
+        if (it != creators_.end()) {
+            return it->second();
+        }
+        return nullptr;
+    }
+
+    std::vector<std::string> registered_types() const {
+        std::vector<std::string> types;
+        types.reserve(creators_.size());
+        for (auto& [name, _] : creators_) {
+            types.push_back(name);
+        }
+        return types;
+    }
+
+private:
+    EffectFactory() = default;
+    std::unordered_map<std::string, Creator> creators_;
+};
+
+/**
+ * RAII helper that registers an effect type at static-init time.
+ * Usage (in .cpp file):
+ *   static EffectRegistrar<MyEffect> reg("My Effect");
+ */
+template <typename T>
+struct EffectRegistrar {
+    explicit EffectRegistrar(const std::string& type_name) {
+        EffectFactory::instance().register_effect(type_name, []() {
+            return std::make_shared<T>();
+        });
+    }
+};
+
+} // namespace GuitarAmp
