+++ src/audio/effects/amp_simulator.cpp
@@ -1,6 +1,9 @@
 #include "audio/effects/amp_simulator.h"
+#include "audio/effect_factory.h"
 
 namespace GuitarAmp {
+
+static EffectRegistrar<AmpSimulator> reg("Amp Sim");
 
 // ============================================================
 // Factory amp model library
@@ -9,96 +12,56 @@ namespace GuitarAmp {
 const std::vector<AmpModel>& get_amp_models() {
     static const std::vector<AmpModel> models = {
         // ---------------------------------------------------------
-        // Clean American — Fender Twin/Deluxe
-        // Scooped mid Baxandall-style tone stack, gentle soft clipping,
-        // glassy highs, warm lows. Designed for sparkling cleans.
+        // Clean American -- Fender Twin/Deluxe
         // ---------------------------------------------------------
         {
             "Clean American",
             "Fender Twin / Deluxe",
             "Sparkling clean, scooped mids, glassy highs",
-            // Tone stack: bass boost, slight mid scoop, treble sparkle
-            200.0f, 3.0f, 0.7f,     // low shelf: 200Hz +3dB
-            800.0f, -2.0f, 0.8f,    // mid peak:  800Hz -2dB (scoop)
-            3500.0f, 2.5f, 0.7f,    // high shelf: 3.5kHz +2.5dB
-            // Saturation: very low gain, pure soft clip, symmetric
-            1.2f,                     // preamp_gain
-            0.0f,                     // saturation_mix (100% soft clip)
-            1.0f,                     // asymmetry (symmetric)
-            0.85f,                    // output_level
-            // Dynamics: slow attack, medium release, no sag
-            0.01f,                    // attack_coeff
-            0.005f,                   // release_coeff
-            0.0f,                     // sag_amount
+            200.0f, 3.0f, 0.7f,
+            800.0f, -2.0f, 0.8f,
+            3500.0f, 2.5f, 0.7f,
+            1.2f, 0.0f, 1.0f, 0.85f,
+            0.01f, 0.005f, 0.0f,
         },
         // ---------------------------------------------------------
-        // British Crunch — Marshall JCM800
-        // Mid-forward tone stack, warm breakup with moderate gain,
-        // asymmetric tube-like saturation. Classic rock crunch.
+        // British Crunch -- Marshall JCM800
         // ---------------------------------------------------------
         {
             "British Crunch",
             "Marshall JCM800",
             "Warm breakup, mid-forward, classic rock crunch",
-            // Tone stack: tight low end, mid push, smooth treble
-            180.0f, -1.0f, 0.8f,    // low shelf: 180Hz -1dB (tight)
-            650.0f, 4.0f, 1.2f,     // mid peak:  650Hz +4dB (mid push)
-            3000.0f, 1.5f, 0.7f,    // high shelf: 3kHz +1.5dB
-            // Saturation: moderate gain, mostly soft clip, asymmetric tubes
-            3.5f,                     // preamp_gain
-            0.15f,                    // saturation_mix (85% soft, 15% hard)
-            0.8f,                     // asymmetry (tube-like asymmetry)
-            0.7f,                     // output_level
-            // Dynamics: medium attack, medium release, slight sag
-            0.05f,                    // attack_coeff
-            0.008f,                   // release_coeff
-            0.15f,                    // sag_amount
+            180.0f, -1.0f, 0.8f,
+            650.0f, 4.0f, 1.2f,
+            3000.0f, 1.5f, 0.7f,
+            3.5f, 0.15f, 0.8f, 0.7f,
+            0.05f, 0.008f, 0.15f,
         },
         // ---------------------------------------------------------
-        // High Gain Modern — Mesa Boogie Rectifier
-        // Tight low-end, scooped mids, aggressive hard clipping.
-        // Dense harmonic saturation for modern metal.
+        // High Gain Modern -- Mesa Boogie Rectifier
         // ---------------------------------------------------------
         {
             "High Gain Modern",
             "Mesa Boogie Rectifier",
             "Tight low-end, scooped mids, aggressive distortion",
-            // Tone stack: tight bass, deep mid scoop, present highs
-            150.0f, 1.5f, 1.0f,     // low shelf: 150Hz +1.5dB (tight)
-            500.0f, -4.5f, 0.9f,    // mid peak:  500Hz -4.5dB (scoop)
-            4000.0f, 2.0f, 0.8f,    // high shelf: 4kHz +2dB
-            // Saturation: high gain, hard clip dominant, slight asymmetry
-            8.0f,                     // preamp_gain
-            0.7f,                     // saturation_mix (30% soft, 70% hard)
-            0.9f,                     // asymmetry
-            0.55f,                    // output_level (lower to compensate gain)
-            // Dynamics: fast attack, slow release, moderate sag
-            0.15f,                    // attack_coeff
-            0.003f,                   // release_coeff
-            0.25f,                    // sag_amount
+            150.0f, 1.5f, 1.0f,
+            500.0f, -4.5f, 0.9f,
+            4000.0f, 2.0f, 0.8f,
+            8.0f, 0.7f, 0.9f, 0.55f,
+            0.15f, 0.003f, 0.25f,
         },
         // ---------------------------------------------------------
-        // Jazz Warm — Roland JC-120
-        // Flat clean, warm highs rolloff, no breakup.
-        // Designed for warm, round jazz tones.
+        // Jazz Warm -- Roland JC-120
         // ---------------------------------------------------------
         {
             "Jazz Warm",
             "Roland JC-120",
             "Flat clean, warm highs rolloff, round tone",
-            // Tone stack: slight bass warmth, flat mids, rolled-off treble
-            250.0f, 2.0f, 0.6f,     // low shelf: 250Hz +2dB (warmth)
-            700.0f, 0.5f, 0.7f,     // mid peak:  700Hz +0.5dB (flat-ish)
-            2800.0f, -3.5f, 0.6f,   // high shelf: 2.8kHz -3.5dB (rolloff)
-            // Saturation: minimal gain, soft clip only, symmetric (solid state)
-            1.0f,                     // preamp_gain
-            0.0f,                     // saturation_mix (100% soft clip)
-            1.0f,                     // asymmetry (symmetric — solid state)
-            0.9f,                     // output_level
-            // Dynamics: very slow attack, slow release, no sag
-            0.005f,                   // attack_coeff
-            0.003f,                   // release_coeff
-            0.0f,                     // sag_amount
+            250.0f, 2.0f, 0.6f,
+            700.0f, 0.5f, 0.7f,
+            2800.0f, -3.5f, 0.6f,
+            1.0f, 0.0f, 1.0f, 0.9f,
+            0.005f, 0.003f, 0.0f,
         },
     };
     return models;
@@ -127,53 +90,6 @@ void AmpSimulator::set_sample_rate(int sample_rate) {
     recompute_coefficients_if_dirty();
 }
 
-void AmpSimulator::compute_low_shelf(float freq, float gain_db, float q) {
-    float A = std::pow(10.0f, gain_db / 40.0f);
-    float w0 = TWO_PI * freq / sample_rate_;
-    float cos_w0 = std::cos(w0);
-    float sin_w0 = std::sin(w0);
-    float alpha = sin_w0 / (2.0f * q);
-    float sqA = std::sqrt(A);
-
-    float a0 = (A + 1) + (A - 1) * cos_w0 + 2 * sqA * alpha;
-    low_shelf_.b0 = (A * ((A + 1) - (A - 1) * cos_w0 + 2 * sqA * alpha)) / a0;
-    low_shelf_.b1 = (2 * A * ((A - 1) - (A + 1) * cos_w0)) / a0;
-    low_shelf_.b2 = (A * ((A + 1) - (A - 1) * cos_w0 - 2 * sqA * alpha)) / a0;
-    low_shelf_.a1 = (-2 * ((A - 1) + (A + 1) * cos_w0)) / a0;
-    low_shelf_.a2 = ((A + 1) + (A - 1) * cos_w0 - 2 * sqA * alpha) / a0;
-}
-
-void AmpSimulator::compute_peaking(float freq, float gain_db, float q) {
-    float A = std::pow(10.0f, gain_db / 40.0f);
-    float w0 = TWO_PI * freq / sample_rate_;
-    float cos_w0 = std::cos(w0);
-    float sin_w0 = std::sin(w0);
-    float alpha = sin_w0 / (2.0f * q);
-
-    float a0 = 1 + alpha / A;
-    mid_peak_.b0 = (1 + alpha * A) / a0;
-    mid_peak_.b1 = (-2 * cos_w0) / a0;
-    mid_peak_.b2 = (1 - alpha * A) / a0;
-    mid_peak_.a1 = (-2 * cos_w0) / a0;
-    mid_peak_.a2 = (1 - alpha / A) / a0;
-}
-
-void AmpSimulator::compute_high_shelf(float freq, float gain_db, float q) {
-    float A = std::pow(10.0f, gain_db / 40.0f);
-    float w0 = TWO_PI * freq / sample_rate_;
-    float cos_w0 = std::cos(w0);
-    float sin_w0 = std::sin(w0);
-    float alpha = sin_w0 / (2.0f * q);
-    float sqA = std::sqrt(A);
-
-    float a0 = (A + 1) - (A - 1) * cos_w0 + 2 * sqA * alpha;
-    high_shelf_.b0 = (A * ((A + 1) + (A - 1) * cos_w0 + 2 * sqA * alpha)) / a0;
-    high_shelf_.b1 = (-2 * A * ((A - 1) + (A + 1) * cos_w0)) / a0;
-    high_shelf_.b2 = (A * ((A + 1) + (A - 1) * cos_w0 - 2 * sqA * alpha)) / a0;
-    high_shelf_.a1 = (2 * ((A - 1) - (A + 1) * cos_w0)) / a0;
-    high_shelf_.a2 = ((A + 1) - (A - 1) * cos_w0 - 2 * sqA * alpha) / a0;
-}
-
 void AmpSimulator::recompute_coefficients_if_dirty() {
     int model_idx = clamp(static_cast<int>(params_[0].value + 0.5f),
                           0, static_cast<int>(get_amp_models().size()) - 1);
@@ -189,9 +105,9 @@ void AmpSimulator::recompute_coefficients_if_dirty() {
         gain_knob != cached_gain_) {
 
         const AmpModel& model = get_amp_models()[model_idx];
-        compute_low_shelf(model.bass_freq, model.bass_gain_db + bass_trim, model.bass_q);
-        compute_peaking(model.mid_freq, model.mid_gain_db + mid_trim, model.mid_q);
-        compute_high_shelf(model.treble_freq, model.treble_gain_db + treble_trim, model.treble_q);
+        low_shelf_.set_low_shelf(model.bass_freq, model.bass_gain_db + bass_trim, model.bass_q, sample_rate_);
+        mid_peak_.set_peaking(model.mid_freq, model.mid_gain_db + mid_trim, model.mid_q, sample_rate_);
+        high_shelf_.set_high_shelf(model.treble_freq, model.treble_gain_db + treble_trim, model.treble_q, sample_rate_);
 
         cached_model_index_ = model_idx;
         cached_bass_ = bass_trim;
@@ -213,7 +129,7 @@ void AmpSimulator::process(float* buffer, int num_samples) {
     float gain_knob = params_[1].value;
     float level = params_[5].value;
 
-    // Effective preamp gain: model base * user gain control (0–2x range)
+    // Effective preamp gain: model base * user gain control (0-2x range)
     float effective_gain = model.preamp_gain * (0.2f + gain_knob * 1.8f);
     float sat_mix = model.saturation_mix;
     float asym = model.asymmetry;
@@ -262,9 +178,7 @@ void AmpSimulator::process(float* buffer, int num_samples) {
         x = soft * (1.0f - sat_mix) + hard * sat_mix;
 
         // --- DC blocking high-pass filter ---
-        float hp_coeff = 0.005f;
-        hp_state_ += hp_coeff * (x - hp_state_);
-        x -= hp_state_;
+        x = dc_block_.hp(x, 0.005f);
 
         // --- Output level ---
         x *= model_output * level;
@@ -282,7 +196,7 @@ void AmpSimulator::reset() {
     mid_peak_.reset();
     high_shelf_.reset();
     envelope_ = 0.0f;
-    hp_state_ = 0.0f;
+    dc_block_.reset();
 }
 
 } // namespace GuitarAmp
