+++ src/audio/effects/compressor.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "audio/effect.h"
+#include "audio/dsp/envelope_follower.h"
 
 namespace GuitarAmp {
 
@@ -14,7 +15,7 @@ public:
 
 private:
     std::vector<EffectParam> params_;
-    float envelope_ = 0.0f;
+    EnvelopeFollower env_;
 };
 
 } // namespace GuitarAmp
