+++ src/audio/effects/distortion.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "audio/effect.h"
+#include "audio/dsp/biquad.h"
 
 namespace GuitarAmp {
 
@@ -14,7 +15,7 @@ public:
 
 private:
     std::vector<EffectParam> params_;
-    float lp_state_ = 0.0f;
+    OnePole tone_lp_;
 };
 
 } // namespace GuitarAmp
