+++ src/audio/effects/reverb.cpp
@@ -1,6 +1,9 @@
 #include "audio/effects/reverb.h"
+#include "audio/effect_factory.h"
 
 namespace GuitarAmp {
+
+static EffectRegistrar<Reverb> reg("Reverb");
 
 // Comb filter delay lengths (in samples at 44100Hz, will be scaled)
 static const int COMB_LENGTHS[] = {1116, 1188, 1277, 1356};
